+++ RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+#### 0.1.2 - 09/01/2016
+* FIXED: package name with dash symbols
+* FIXED: replaces variable expressions
+
 #### 0.1.1 - 09/01/2016
 * Setup project
 * basic updater